+++ src/client.ts
@@ -325,7 +325,8 @@ export class Client {
       const opts: RequestApiOptions = {
         method: 'get',
         host: this.config.apiHost,
-        authToken: this.config.authToken
+        authToken: this.config.authToken,
+        logger: this.logger
       };
       jidPromise = this.http.requestApiWithRetry('users/me', opts).promise
         .then(res => res.body.chat.jabberId);