+++ src/webrtc.ts
@@ -1,21 +1,23 @@
-import { definitions, Propose } from './stanza-definitions/webrtc-signaling';
 import { EventEmitter } from 'events';
-import { ReceivedMessage } from 'stanza/protocol';
+import { ExternalService, ExternalServiceList, ReceivedMessage } from 'stanza/protocol';
 import { toBare } from 'stanza/JID';
-import { GenesysCloudMediaSession, SessionEvents, SessionType } from './types/media-session';
 import LRU from 'lru-cache';
 import { JingleAction } from 'stanza/Constants';
 import { SessionManager } from 'stanza/jingle';
 import { v4 } from 'uuid';
 import { Jingle } from 'stanza';
-import { isAcdJid, isScreenRecordingJid, isSoftphoneJid, isVideoJid, calculatePayloadSize } from './utils';
 import { StatsEvent } from 'webrtc-stats-gatherer';
 import throttle from 'lodash.throttle';
+import JingleSession from 'stanza/jingle/Session';
+import { isFirefox } from 'browserama';
+
+import { definitions, Propose } from './stanza-definitions/webrtc-signaling';
+import { GenesysCloudMediaSession, SessionEvents, SessionType } from './types/media-session';
+import { isAcdJid, isScreenRecordingJid, isSoftphoneJid, isVideoJid, calculatePayloadSize, retryPromise, RetryPromise } from './utils';
 import Client from '.';
 import { formatStatsEvent } from './stats-formatter';
 import { ClientOptions } from './client';
-import JingleSession from 'stanza/jingle/Session';
-import { isFirefox } from 'browserama';
+import { ExtendedRTCIceServer } from './types/interfaces';
 
 const events = {
   REQUEST_WEBRTC_DUMP: 'requestWebrtcDump', // dump triggered by someone in room
@@ -41,6 +43,8 @@ const events = {
 };
 
 const desiredMaxStatsSize = 15000;
+const MAX_DISCO_RETRIES = 2;
+const ICE_SERVER_TIMEOUT = 15000; // 15 seconds
 
 type ProposeStanza = ReceivedMessage & { propose: Propose };
 
@@ -71,6 +75,8 @@ export class WebrtcExtension extends EventEmitter {
   private statsSizeDecreaseAmount = 3000;
   private statBuffer = 0;
   private throttledSendStats: any;
+  private discoRetries = 0;
+  private refreshIceServersRetryPromise: RetryPromise<ExtendedRTCIceServer[]> | undefined;
 
   get jid (): string {
     return this.client._stanzaio.jid;
@@ -102,6 +108,14 @@ export class WebrtcExtension extends EventEmitter {
     Object.assign(this.client._stanzaio.jingle.config.peerConnectionConfig, {
       sdpSemantics: 'unified-plan'
     });
+
+    /* clear out stanzas default use of google's stun server */
+    this.client._stanzaio.jingle.config.iceServers = [];
+    /**
+     * NOTE: calling this here should not interfere with the `webrtc.ts` extension
+     *  refreshingIceServers since that is async and this constructor is sync
+     */
+    this.setIceServers([]);
   }
 
   prepareSession (options: any) {
@@ -215,8 +229,14 @@ export class WebrtcExtension extends EventEmitter {
   }
 
   addEventListeners () {
-    this.client.on('connected', async () => {
-      await this.refreshIceServers();
+    this.client.on('connected', () => {
+      return this.refreshIceServers()
+        .catch((error) =>
+          this.logger.error('Error fetching ice servers after streaming-client connected', {
+            error,
+            channelId: this.client.config.channelId
+          })
+        );
     });
 
     this.client._stanzaio.jingle.on('log', (level, message, data?) => {
@@ -508,37 +528,86 @@ export class WebrtcExtension extends EventEmitter {
     this.logger.info('sent jingle retract', { sessionId, conversationId: session.propose.conversationId });
   }
 
-  async refreshIceServers (): Promise<any[]> {
+  async refreshIceServers (): Promise<ExtendedRTCIceServer[]> {
+    if (!this.refreshIceServersRetryPromise) {
+      this.refreshIceServersRetryPromise = retryPromise<ExtendedRTCIceServer[]>(
+        this._refreshIceServers.bind(this),
+        (error): boolean => {
+          if (++this.discoRetries > MAX_DISCO_RETRIES) {
+            this.logger.warn('fetching ice servers failed. max retries reached.', {
+              retryAttempt: this.discoRetries,
+              MAX_DISCO_RETRIES,
+              error,
+              channelId: this.client.config.channelId
+            });
+            return false;
+          }
+
+          this.logger.warn('fetching ice servers failed. retrying', {
+            retryAttempt: this.discoRetries,
+            error,
+            channelId: this.client.config.channelId
+          });
+          return true;
+        },
+        0,
+        this.client.logger
+      );
+    }
+
+    return this.refreshIceServersRetryPromise.promise
+      .finally(() => {
+        this.discoRetries = 0;
+        this.refreshIceServersRetryPromise = undefined;
+      });
+  }
+
+  async _refreshIceServers (): Promise<ExtendedRTCIceServer[]> {
     const server = this.client._stanzaio.config.server;
     const turnServersPromise = this.client._stanzaio.getServices(
-      server as any,
+      server as string,
       'turn',
       '1'
     );
     const stunServersPromise = this.client._stanzaio.getServices(
-      server as any,
+      server as string,
       'stun',
       '1'
     );
 
-    const [turnServers, stunServers] = await Promise.all([
-      turnServersPromise,
-      stunServersPromise
-    ]);
+    const servicesPromise = new Promise<[Required<ExternalServiceList>, Required<ExternalServiceList>]>((resolve, reject) => {
+      setTimeout(() => {
+        reject(new Error('Timeout waiting for refresh ice servers to finish'));
+      }, ICE_SERVER_TIMEOUT);
+
+      Promise.all([
+        turnServersPromise,
+        stunServersPromise
+      ])
+        .then(([turn, stun]) => {
+          resolve([turn, stun] as [Required<ExternalServiceList>, Required<ExternalServiceList>]);
+        })
+        .catch(reject);
+    });
+
+    const [turnServers, stunServers] = await servicesPromise;
+    // const [turnServers, stunServers] = await Promise.all([turnServersPromise, stunServersPromise]) as any;
+
     this.logger.debug('STUN/TURN server discovery result', {
       turnServers,
       stunServers
     });
+
     const iceServers = [
-      ...(turnServers.services as any),
-      ...(stunServers.services as any)
-    ].map((service) => {
+      ...turnServers.services,
+      ...stunServers.services
+    ].map((service: ExternalService) => {
       const port = service.port ? `:${service.port}` : '';
-      const ice: RTCIceServer & { type: string } = {
-        type: service.type,
+      const ice: ExtendedRTCIceServer = {
+        type: service.type as string,
         urls: `${service.type}:${service.host}${port}`
       };
-      if (['turn', 'turns'].includes(service.type)) {
+      if (['turn', 'turns'].includes(service.type as string)) {
         if (service.transport && service.transport !== 'udp') {
           ice.urls += `?transport=${service.transport}`;
         }
@@ -553,7 +622,7 @@ export class WebrtcExtension extends EventEmitter {
     });
 
     this.setIceServers(iceServers);
-    if (!stunServers.services!.length) {
+    if (!stunServers.services.length) {
       this.logger.info('No stun servers received, setting iceTransportPolicy to "relay"');
       this.setIceTransportPolicy('relay');
     } else {