+++ test/unit/webrtc.test.ts
@@ -5,17 +5,15 @@ import { Agent, createClient } from 'stanza';
 import { JingleAction } from 'stanza/Constants';
 import { v4 } from 'uuid';
 import { EventEmitter } from 'events';
+import browserama from 'browserama';
 
 import { WebrtcExtension } from '../../src/webrtc';
 import { GenesysCloudMediaSession } from '../../src/types/media-session';
 import * as utils from '../../src/utils';
-import { wait } from '../helpers/testing-utils';
 import * as statsFormatter from '../../src/stats-formatter';
 import { HttpClient } from '../../src/http-client';
-import browserama from 'browserama';
 
 jest.mock('../../src/types/media-session');
-
 class Client extends WildEmitter {
   connected = false;
   emit!: (event: string, ...data: any) => void;
@@ -29,7 +27,7 @@ class Client extends WildEmitter {
   _stanzaio: Agent;
   http: HttpClient;
 
-  constructor(public config: any) {
+  constructor (public config: any) {
     super();
     this._stanzaio = createClient({});
     this.http = new HttpClient();
@@ -44,6 +42,10 @@ function shimCreatePeerConnection (client) {
   client._stanzaio.jingle.createPeerConnection = () => {
     return new FakePeerConnection();
   };
+}
+
+function flush () {
+  return new Promise(res => setImmediate(res));
 }
 
 describe('constructor', () => {
@@ -99,6 +101,25 @@ describe('prepareSession', () => {
 });
 
 describe('addEventListeners', () => {
+  it('should refresh ice servers on "connected"', async () => {
+    const client = new Client({});
+    const webrtc = new WebrtcExtension(client as any, {} as any);
+    const error = new Error('Bad timing');
+
+    const spy = jest.spyOn(webrtc.logger, 'error');
+    jest.spyOn(webrtc, 'refreshIceServers').mockRejectedValue(error);
+    client.config.channelId = 'my-ws-channel';
+
+    client.emit('connected');
+
+    await flush();
+
+    expect(spy).toHaveBeenCalledWith('Error fetching ice servers after streaming-client connected', {
+      error,
+      channelId: client.config.channelId
+    });
+  });
+
   it('should listen for jingle log messages', () => {
     const client = new Client({});
     const webrtc = new WebrtcExtension(client as any, {} as any);
@@ -823,6 +844,94 @@ describe('refreshIceServers', () => {
       { type: 'turn', urls: 'turn:turn.server.com:123', username: 'user1', credential: 'pass1' },
       { type: 'turn', urls: 'turn:turn.server.com:456?transport=tcp', username: 'user2', credential: 'pass2' },
     ]);
+  });
+
+  it('should retry if getting the servers fails', async () => {
+    const client = new Client({});
+    const webrtc = new WebrtcExtension(client as any, {} as any);
+
+    const spy = jest.spyOn(webrtc, 'setIceServers');
+    jest.spyOn(client._stanzaio, 'getServices')
+      /* signifies the first call */
+      .mockRejectedValueOnce(new Error('Failed to fetch servers'))
+      .mockRejectedValueOnce(new Error('Failed to fetch servers'))
+      /* signifies the second call */
+      .mockReturnValueOnce({
+        services: [
+          { type: 'turn', host: 'turn.server.com' },
+          { port: 123, type: 'turn', host: 'turn.server.com', username: 'user1', password: 'pass1' },
+          { port: 456, type: 'turn', host: 'turn.server.com', username: 'user2', password: 'pass2', transport: 'tcp' }
+        ]
+      } as any)
+      .mockReturnValueOnce({
+        services: [
+          { port: 234, type: 'stun', host: 'turn.server.com' }
+        ]
+      } as any);
+
+    await webrtc.refreshIceServers();
+
+    expect(spy).toHaveBeenCalledWith([
+      { type: 'turn', urls: 'turn:turn.server.com' },
+      { type: 'turn', urls: 'turn:turn.server.com:123', username: 'user1', credential: 'pass1' },
+      { type: 'turn', urls: 'turn:turn.server.com:456?transport=tcp', username: 'user2', credential: 'pass2' },
+      { type: 'stun', urls: 'stun:turn.server.com:234' }
+    ]);
+  });
+
+  it('should try 3 times and then fail', async () => {
+    const client = new Client({});
+    const webrtc = new WebrtcExtension(client as any, {} as any);
+
+    const spy = jest.spyOn(client._stanzaio, 'getServices')
+      /* signifies all calls fail */
+      .mockRejectedValue(new Error('Failed to fetch servers'));
+
+    try {
+      await webrtc.refreshIceServers();
+      fail('should have not been able to fetch ice servers');
+    } catch (error) {
+      expect(spy).toHaveBeenCalledTimes(3 * 2); // 3 times total but each call calls `getServices` twice
+      /* resets state */
+      expect(webrtc['discoRetries']).toBe(0);
+      expect(webrtc['refreshIceServersRetryPromise']).toBe(undefined);
+    }
+  });
+
+  it('should timeout after 15 seconds', async () => {
+    jest.useFakeTimers();
+
+    const client = new Client({});
+    const webrtc = new WebrtcExtension(client as any, {} as any);
+
+    const spy = jest.spyOn(client._stanzaio, 'getServices')
+      /* have the call to fetch services "hang" */
+      .mockImplementation(() => new Promise(res => setTimeout(res, 100 * 1000)));
+
+    const promise = webrtc.refreshIceServers();
+
+    expect(spy).toHaveBeenCalledTimes(1 * 2);
+
+    /* I don't know why these timers & flushes have to work this way... but it is the only combo that works */
+    jest.advanceTimersByTime(15000);
+    await flush();
+
+    jest.advanceTimersByTime(15000);
+    await flush();
+
+    expect(spy).toHaveBeenCalledTimes(2 * 2);
+
+    jest.advanceTimersByTime(15000);
+    expect(spy).toHaveBeenCalledTimes(3 * 2);
+
+    try {
+      await promise;
+      fail('ice servers should have timedout');
+    } catch (error) {
+      expect(error.message).toBe('Timeout waiting for refresh ice servers to finish');
+    }
+
+    jest.clearAllTimers();
   });
 });